fix(jwt-token): Fixed the implementation of token service
