+++ backend/src/application/services/token/token-manager-service.ts
@@ -1,0 +1,9 @@
+export interface StudentTokenManagerGenerateInput {
+  studentId: string;
+}
+
+export interface StudentTokenManager {
+  generate: (data: StudentTokenManagerGenerateInput) => Promise<string>;
+  verify: (token: string) => Promise<boolean>;
+  decrypt: (token: string) => Promise<object>;
+}
