--- backend/src/application/services/token/token-manager.ts
@@ -1,5 +1,0 @@
-export interface StudentTokenManager {
-  generate: (data: object) => Promise<string>;
-  verify: (token: string) => Promise<boolean>;
-  decrypt: (token: string) => Promise<object>;
-}
